# projects/control-service/projects/model/apidefs/datajob-api/api.yaml
...
openapi: 3.0.3
servers:
  - url: http://localhost:8092
    description: quickstart-vdk default local setup
info:
  title: Versatile Data Kit Control Service API
  version: '1.0'
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  description: "The Data Jobs API of Versatile Data Kit Control Service.
    Data Jobs allows Data Engineers to implement automated pull ingestion (E in ELT)
    and batch data transformation into a database (T in ELT).
    See also https://github.com/vmware/versatile-data-kit/wiki/Introduction

    The API has resource-oriented URLs, JSON-encoded responses, and uses standard HTTP response codes, authentication, and verbs.
    The API enables creating, deploying, managing and executing Data Jobs in the runtime environment.<br>
    <br>
    ![](https://github.com/vmware/versatile-data-kit/wiki/vdk-data-job-lifecycle-state-diagram.png)
    <br>
    The API reflects the usual Data Job Development lifecycle:<br>
    <li> Create a new data job (webhook to further configure the job, e.g authorize its creation, setup permissions, etc).
    <li> Download keytab. Develop and run the data job locally.
    <li> Deploy the data job in cloud runtime environment to run on a scheduled basis.
    <br><br>
    If Authentication is enabled, pass OAuth2 access token in HTTP header 'Authorization: Bearer [access-token-here]' (https://datatracker.ietf.org/doc/html/rfc6750).
    <br

...
    <br><br>
    <b>API Evolution</b><br>
...
paths:
...
  '/data-jobs/for-team/{team_name}/jobs':
    get:
      operationId: jobsQuery
      summary: Query Data Jobs details using GraphQL
      tags:
        - Data Jobs
      description: |
        List data Jobs with GraphQL like query. By choosing which field to be returned you can control the output.
        You can learn more about the GraphQL queries by visiting [GraphQL official website](https://graphql.org/learn/queries/)
        Query should be provided as GET parameter, not by POST body. Don't worry about the spaces and tabs
        Keep in mind that each aditional field <b>could make query response time slower, for instance deployments</b>,
        it's best if you request only what you need <br/><br/>

        The <b>pageNumber</b> and <b>pageSize</b> arguments are required! Page number should be a <b>number greater than 1</b>,
        and pageSize <b>should be greater than 1</b> (per page)
        Check the latest example for the full list of supported query fields.<br/><br/>

        Simplest query that you could make is to fetch the job names
        ```
        {
          jobs(pageNumber: 1, pageSize: 25) {
            content {
              jobName
            }
          }
        }
        ```

        You could also use filtering and sorting function. Filter object has <b>property, pattern and sort</b> fields.<br/>
        * <b>property</b> points out which field you want to filter, if you point out some other field that is not supported,
        an invalid response will be return.<br/>
        * <b>pattern</b> should be a non-empty string which the provided property should contains [ignoring cases], for instance: `starshot` pattern will match
        <b>import-starshot-sql, StarShot-servers and notify-starshot</b> job names, but it won't match <b>stars-shot-daily-prune`</b>
        If a pattern string is not provided, then you must atleast provide the property field<br/>
        * <b>sort</b> should be an enum value - ASC (ascending) or DESC (descending) option [not required, default is ASC]
        Multiple filters could be applied, but <b>maximum one should contain sorting</b>!
        ```
        {
          jobs(
            pageNumber: 1,
            pageSize: 25,
            filter: [{
              property: "jobName",
              pattern: "starshot",
              sort: DESC
            }],
          ) {
            content {
              jobName
            }
          }
        }
        ```

        You could also search for a string into the properties that you are requesting, for instance:
        This query will search for job names, team names and descriptions which contains the provided "starshot" string
        ```
        {
          jobs(
            pageNumber: 1,
            pageSize: 25,
            search: "starshot"
          ) {
            content {
              jobName,
              config {
                team
                description
              }
            }
          }
        }
        ```

        Data jobs execution could also be searched by providing arguments to the <b>execution</b> field.
        Same as parent query arguments, the <b>pageNumber</b> and <b>pageSize</b> arguments are required! Page number should be a <b>number greater than 1</b>,
        and pageSize <b>should be between 1 and 100 results</b> (per page). You can also <b>filter</b> using the similar object structure as the parent query,
        but currently <b>filtering is not supported</b>, you can only provide field for sorting.
        This query will search
        ```
        {
          jobs(
            pageNumber: 1,
            pageSize: 25,
          ) {
            content {
              jobName,
              deployments {
                id
                executions(
                  pageNumber: 1,
                  pageSize: 5,
                  filter: [{
                    teamNameIn: ["starshot"]
                  }],
                  order: {
                    property: "startTime",
                      direction: DESC
                  }
                ) {
                  id
                  status
                  startTime
                  endTime
                {
              }
            }
          }
        }
        ```

        Full example of currently available for fetching fields. Note that if you combine searching and filtering, first
        it will apply filters and then within filtered jobs it will apply the search, vice versa is currently not supported:
        ```
        {
          jobs(
            pageNumber: 1,
            pageSize: 25,
            search: "daily",
            filter: [{
              property: "jobName",
              pattern: "import-sql",
            },{
              property: "team",
              pattern: "starshot",
              sort: DESC
            },{
              property: "deployments.enabled",
              pattern: "enabled",
            }],
          ) {
            content {
              jobName
              config {
                team
                description
                sourceUrl
                schedule {
                  scheduleCron
                  nextRunEpochSeconds
                }
                contacts {
                  notifiedOnJobFailureUserError
                  notifiedOnJobFailurePlatformError
                  notifiedOnJobSuccess
                  notifiedOnJobDeploy
                }
              }
              deployments {
                id
                enabled
                jobVersion
                mode
                executions(
                  pageNumber: 1,
                  pageSize: 25,
                  filter: [{
                    teamNameIn: ["starshot"]
                  }],
                  order: {
                    property: "startTime",
                      direction: DESC
                  }
                ) {
                  id
                  type
                  status
                  message
                  startTime
                  endTime
                  opId
                  vkdVersion
                  jobVersion
                  jobSchedule
                  resourcesCpuRequest
                  resourcesCpuLimit
                  resourcesMemoryRequest
                  resourcesMemoryLimit
                  deployedDate
                  deployedBy
                  startedBy
                  logsUrl
                }
              }
            }
            totalPages
            totalItems
            }
          }
        }
        ```

      parameters:
        - name: team_name
          description: The Team which owns the Data Job
          schema:
            type: string
          in: path
          required: true
        - name: query
          description: Request a GraphQL-like query.
          schema:
            type: string
            example: "{
                jobs(pageNumber: 1, pageSize: 20, filter: []) {
                  content {
                    jobName
                    config {
                      team
                      description
                      schedule {
                        scheduleCron
                        nextRunEpochSeconds
                      }
                      sourceUrl
                      contacts {
                        notifiedOnJobFailureUserError
                        notifiedOnJobFailurePlatformError
                        notifiedOnJobSuccess
                        notifiedOnJobDeploy
                      }
                    }
                  }
                  totalPages
                  totalItems
                }
              }"
            default: "{
                jobs(pageNumber: 1, pageSize: 20, filter: []) {
                  content {
                    jobName
                    config {
                      team
                      description
                      schedule {
                        scheduleCron
                      }
                    }
                  }
                  totalPages
                  totalItems
                }
              }"
          required: false
          in: query
        - name: operation_name
          description: If the query is specified and contains several named operations, an operation_name query parameter can be used to control which one should be executed
          schema:
            type: string
            example: jobs
          required: false
          in: query
        - name: variables
          description: If query is specified then variables can be sent as a JSON-encoded string in an additional query parameter called variables
          schema:
            type: string
            example: |
              {
                "pageSize": "10",
                "pageNumber": "1"
              }
          required: false
          in: query
      responses:
        '200':
          $ref: '#/components/responses/DataJobQueryResponse'
        '400':
          $ref: '#/components/responses/DataJobQueryResponseWithError'
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DataJob'
        required: true
      parameters:
        - name: team_name
          description: The Team which owns the Data Job
          schema:
            type: string
          in: path
          required: true
        - in: query
          name: name
          schema:
            type: string
          required: false
          description: The Name of the Data Job
      tags:
        - Data Jobs
      responses:
        '201':
          description: Created successfully.
        '400':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
      operationId: dataJobCreate
      summary: Creates a new Data Job | (Stable)
      description: |
        <i>(Introduced in v1.0)</i> | Creates a new Data Job.<br>
        The name of the Data job should comply to the following naming convention:<br>
        <li>should start with a letter
        <li>50 characters at most
        <li>5 characters at least
        <li>should contain only lowercase alphanumeric symbols and dash [a-z0-9\-]
        <br><br>
        When succesfully created clients can download the keytab associated with the Data Job to run it locally.<br>
        To deploy so that it can be executed reguarly in the runtime environment use /data-jobs/for-team/{team_name}/jobs/{job_name}/deployments API.
  '/data-jobs/for-team/{team_name}/jobs/{job_name}':
    summary: |
      A Data Job is a sequence of Python and/or SQL scripts and configurations, executed by Versatile Data Kit SDK.<br>
      See more in https://github.com/vmware/versatile-data-kit/wiki/Getting-Started
    get:
      tags:
        - Data Jobs
      responses:
        '200':
          $ref: '#/components/responses/DataJob'
        '404':
          $ref: '#/components/responses/Error'
      operationId: dataJobRead
      summary: Retrieves details of an existing Data Job by specifying the name of the Data Job. | (Stable)
      description: <i>(Introduced in v1.0)</i>
    put:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DataJob'
        required: true
      tags:
        - Data Jobs
      responses:
        '201':
          description: Created successfully.
        '204':
          description: Updated successfully.
        '400':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
      operationId: dataJobUpdate
      summary: Update Data Job. | (Stable)
      description: <i>(Introduced in v1.0)</i>
    delete:
      tags:
        - Data Jobs
      responses:
        '200':
          description: Deleted successfully.
        '404':
          $ref: '#/components/responses/Error'
      operationId: dataJobDelete
      summary: Delete Data Job | (Stable)
      description: |
        <i>(Introduced in v1.0)</i> | Delete Data Job including its state (properties) and deployments.
        Currently executing Data Jobs will be left to finish.
    parameters:
      - name: team_name
        description: Team Name
        schema:
          type: string
        in: path
        required: true
      - name: job_name
        description: Data Job Name
        schema:
          type: string
        in: path
        required: true
  '/data-jobs/for-team/{team_name}/jobs/{job_name}/keytab':
    summary: |
      List of credentials to be used by a data job - for example kerberos key to authenticate to the database.
      Note: will be renamed to /credentials as it seems more clear/generic
    get:
      tags:
        - Data Jobs
      responses:
        '200':
          description: Success
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
      operationId: dataJobKeytabDownload
      summary: Get data job keytab. | (Stable)
      description: <i>(Introduced in v1.0)</i>
    parameters:
      - name: team_name
        description: Team Name
        schema:
          type: string
        in: path
        required: true
      - name: job_name
        description: Data Job Name.
        schema:
          type: string
        in: path
        required: true
  '/data-jobs/for-team/{team_name}/jobs/{job_name}/team/{new_team}':
    put:
      tags:
        - Data Jobs
      responses:
        '200':
          description: Team changed successfully.
        '400':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
      operationId: dataJobTeamUpdate
      summary: Update API for Data Jobs team | (Stable)
      description: <i>(Introduced in v1.0)</i> | Update API for Data Jobs team
      parameters:
        -  name: team_name
           description: Current Team Name.
           schema:
             type: string
           in: path
           required: true
        -  name: new_team
           description: New Team Name.
           schema:
             type: string
           in: path
           required: true
        - name: job_name
          description: Data Job Name.
          schema:
            type: string
          in: path
          required: true
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    DataJob:
      description: Data Job Details
      required:
        - job_name
        - description
        - config
      type: object
      properties:
        job_name:
          description: Data Job name
          type: string
          example: starshot-processing-vmc-fact-daily
        description:
          description: Short Data Job description
          type: string
          example: Data Job responsible for transforming vmc-related fact tables on a daily basis.
        config:
          $ref: '#/components/schemas/DataJobConfig'
        team:
          description: (Optional) Team name
          type: string
          example: starshot (Optional property)
    Error:
      description: Contains description for one or more errors detected.
      required:
        - messages
      type: object
      properties:
        messages:
          description: Error messages
          type: array
          items:
            type: string
          example:
            - Data Job starshot-processing-vmc-fact-daily not found
...
    DataJobConfig:
      description: Data Job configuration properties.
      type: object
      properties:
        db_default_type:
          example: TRINO
          description: |
            Default DB connection provided for the job, e.g. 'TRINO', 'IMPALA', 'REDSHIFT'.<br>
          type: string
        contacts:
          $ref: '#/components/schemas/DataJobContacts'
        schedule:
          $ref: '#/components/schemas/DataJobSchedule'
          deprecated: true
        generate_keytab:
          example: false
          description: Enable Disable flag for generating secret with keytab
          type: boolean
          default: true
        enable_execution_notifications:
          example: false
          description: (Optional) Specifies whether to dispatch email notifications per data job execution.
          type: boolean
          default: true
        notification_delay_period_minutes:
          format: int32
          description: |
            (Optional) Specifies the time (in minutes) a job execution is allowed to be delayed from
            its schedule before an alert is triggered.
            The value of this property for each data job is exposed as a Prometheus metric by the service
            at the /data-jobs/debug/prometheus endpoint. These metrics (taurus_datajob_notification_delay)
            are subsequently used in JobDelay Prometheus rules to generate the alerts.
          type: integer
          example: 60
          default: 240
    DataJobSchedule:
      description: Schedule configuration
      type: object
      properties:
        schedule_cron:
          example: 0 0 13 * 5
          description: |
            For format see https://en.wikipedia.org/wiki/Cron<br>
            The cron expression is evaluated in UTC time.
            If it is time for a new job run and the previous job run hasn't finished yet, the cron job kills and replaces the currently running job run with a new job run.
            Jobs configured to run more often than once per hour are not supported and their schedule may be overridden by the platform.
            To distribute load evenly, Administrators may override the minute you specified.
            Use https://crontab.guru for help.
          type: string
    DataJobContacts:
      description: Data Job contacts
      type: object
      properties:
        notified_on_job_failure_user_error:
          description: |
            List of email addresses to be notified on job execution failure caused by user code or user configuration problem.
            E.g. if the job contains a SQL script with a syntax error.
          type: array
          items:
            type: string
          example:
            - auserov@example.mail.com
        notified_on_job_failure_platform_error:
          description: List of email addresses to be notified on job execution failure caused by a platform problem
          type: array
          items:
            type: string
          example:
            - auserov@example.mail.com
        notified_on_job_success:
          description: List of email addresses to be notified on job execution success
          type: array
          items:
            type: string
          example:
            - auserov@example.mail.com
        notified_on_job_deploy:
          description: List of email addresses to be notified of job deployment outcome
          type: array
          items:
            type: string
          example:
            - auserov@example.mail.com
...
    DataJobQueryResponseWithError:
      description: Query response containing Data Jobs
      type: object
      properties:
        errors:
          type: array
          items:
            type: object
          description: Errors while making query (validation errors, exceptions, etc)
          example:
            application/json: [
               {
                 "message": "Validation error of type FieldUndefined",
                 "locations": [],
                 "description": "Field 'someField' in type 'DataJob' is undefined",
                 "validationErrorType": "FieldUndefined",
                 "queryPath": [
                   "jobs",
                   "content",
                   "someField"
                 ],
                 "extensions": null,
                 "errorType": "ValidationError",
                 "path": null
               }
             ]
        data:
          $ref: '#/components/schemas/DataJobPage'
    DataJobQueryResponse:
      description: Query response containing Data Jobs
      type: object
      properties:
        errors:
          type: array
          items:
            type: object
          description: Errors while making query (validation errors, exceptions, etc)
        data:
          $ref: '#/components/schemas/DataJobPage'
    DataJobPage:
      description: Page object containing Data Jobs list with information for total elements and pages
      type: object
      properties:
        content:
          type: array
          items:
            type: object
          example:
            application/json: [
              {
                "jobName": "starshot-processing-vmc-fact-daily",
                "config": {
                  "team": "starshot",
                  "description": "Data Job responsible for transforming vmc related fact tables on daily basis",
                  "schedule": {
                    "scheduleCron": "5 0 * 8 *",
                    "nextRun": 1618914371
                  },
                  "sourceUrl": "https://github.com/product-analytics/data-jobs/tree/master/starshot-processing-vmc-fact-daily",
                  "contacts": {
                    "notifiedOnJobFailureUserError": "[auser@example.mail.com]",
                    "notifiedOnJobFailurePlatformError": "[auser2@example.mail.com, auser@example.mail.com]",
                    "notifiedOnJobSuccess": "[auser@example.mail.com]",
                    "notifiedOnJobDeploy": "[auser2@example.mail.com, auser@example.mail.com]"
                  }
                }
              },
              ...
            ]
        totalItems:
          description: Number of elements which meet the given query requirement
          type: integer
          example: 100
        totalPages:
          description: Number of pages with elements which meet the given query requirement
          type: integer
          example: 5
  responses:
...
    DataJob:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/DataJob'
      description: Data Job details
    Error:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
      description: A response for any errors detected
...
    DataJobQueryResponse:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/DataJobQueryResponse'
      description: Data Job query response
    DataJobQueryResponseWithError:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/DataJobQueryResponseWithError'
      description: Data Job query response with error
security:
  - bearerAuth: []
...